# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.Login:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.Post:
    properties:
      category:
        type: integer
      description:
        type: string
      title:
        type: string
    required:
    - category
    - description
    - title
    type: object
  dto.Token:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  dto.User:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  serializer.CategoryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  serializer.PostResponse:
    properties:
      category:
        $ref: '#/definitions/serializer.CategoryResponse'
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: abdurraufraihan@gmail.com
    name: Abdur Rauf Raihan
    url: https://linkedin.com/abdurraufraihan
  description: A Blog API in Go using Gin framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Blog Api
  version: "1.0"
paths:
  /login:
    post:
      description: Responds with token pair as JSON.
      parameters:
      - description: Login dto
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: User login
      tags:
      - auth
  /posts:
    get:
      description: Responds with the list of all posts as JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.PostResponse'
      summary: Get posts list
      tags:
      - posts
    post:
      description: Responds with post as JSON.
      parameters:
      - description: Post dto
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.Post'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/serializer.PostResponse'
      summary: Insert post
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Responds with post as JSON.
      parameters:
      - description: delete post by id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete post
      tags:
      - posts
    get:
      description: Responds with post as JSON.
      parameters:
      - description: search post by id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.PostResponse'
      summary: Get post
      tags:
      - posts
    put:
      description: Responds with post as JSON.
      parameters:
      - description: update post by id
        in: path
        name: id
        required: true
        type: integer
      - description: Post dto
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.PostResponse'
      summary: Update post
      tags:
      - posts
  /signup:
    post:
      description: Responds with user data as JSON.
      parameters:
      - description: User dto
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: User register
      tags:
      - auth
  /token/refresh:
    post:
      description: Responds with token pair as JSON.
      parameters:
      - description: Token dto
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.Token'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Refresh user token
      tags:
      - auth
  /token/verify:
    post:
      description: Responds with is_valid status as JSON.
      parameters:
      - description: Token dto
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.Token'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Verify user token
      tags:
      - auth
swagger: "2.0"